refactor: add terminating restart policy
This change adds a new restart policy "terminating". This restart policy
is not intended to be set by users. It is used by the wrapper internally
to indicate that a process should be stopped and not restarted. Previous
to this change, the restart policy "never" was used for this purpose,
but that policy has specific behavior, so it is clearer to create a new
policy that indicates that the process is ending.